refactor: centralize core utilities
