+++ src/mcp_server/core/__init__.py
@@ -1,7 +1,18 @@
-"""Core package for MCP server utilities."""
+"""Core package for MCP server utilities.
+
+This package centralizes configuration loading and structured logging helpers
+used throughout the project.
+"""
 
 from __future__ import annotations
 
+from .config import ServerSettings, get_config, get_server_url, load_environment
 from .logging import configure_logging
 
-__all__ = ["configure_logging"]
+__all__ = [
+    "ServerSettings",
+    "get_config",
+    "get_server_url",
+    "load_environment",
+    "configure_logging",
+]
